refactoring projectileemmitersystem

+++ 2DGameEngine/src/Systems/ProjectileEmitterSystem.h
@@ -13,31 +13,39 @@
 #include "../Events/KeyPressedEvent.h"
 #include <SDL2/SDL.h>
 
-
-class ProjectileEmitterSystem: public System {
+/**
+ * 
+ */
+class ProjectileEmitterSystem: public System 
+{
     public:
-        ProjectileEmitterSystem() {
+        ProjectileEmitterSystem() 
+        {
             RequireComponent<ProjectileEmitterComponent>();
             RequireComponent<TransformComponent>();
         }
 
-        void SubscribeToEvents(std::unique_ptr<EventBus>& eventBus) {
+        void SubscribeToEvents(std::unique_ptr<EventBus>& eventBus) 
+        {
             eventBus->SubscribeToEvent<KeyPressedEvent>(this, &ProjectileEmitterSystem::OnKeyPressed);
         }
 
-        void OnKeyPressed(KeyPressedEvent& event) {
+        void OnKeyPressed(KeyPressedEvent& event) 
+        {
             if (event.symbol == SDLK_SPACE) 
             {
                 Logger::Log("SPACE BUTTON");
                 for (auto entity: GetSystemEntity()) {
-                    if (entity.HasComponent<CameraFollowComponent>()) {
+                    if (entity.HasComponent<CameraFollowComponent>()) 
+                    {
                         const auto projectileEmitter = entity.GetComponent<ProjectileEmitterComponent>();
                         const auto transform = entity.GetComponent<TransformComponent>();
                         const auto rigidbody = entity.GetComponent<RigidBodyComponent>();
 
                         // If parent entity has sprite, start the projectile position in the middle of the entity
                         glm::vec2 projectilePosition = transform.position;
-                        if (entity.HasComponent<SpriteComponent>()) {
+                        if (entity.HasComponent<SpriteComponent>()) 
+                        {
                             auto sprite = entity.GetComponent<SpriteComponent>();
                             projectilePosition.x += (transform.scale.x * sprite.width / 2);
                             projectilePosition.y += (transform.scale.y * sprite.height / 2);
@@ -47,10 +55,12 @@ class ProjectileEmitterSystem: public System {
                         glm::vec2 projectileVelocity = projectileEmitter.projectileVelocity;
                         int directionX = 0;
                         int directionY = 0;
+
                         if (rigidbody.velocity.x > 0) directionX = +1;
                         if (rigidbody.velocity.x < 0) directionX = -1;
                         if (rigidbody.velocity.y > 0) directionY = +1;
                         if (rigidbody.velocity.y < 0) directionY = -1;
+                        
                         projectileVelocity.x = projectileEmitter.projectileVelocity.x * directionX;
                         projectileVelocity.y = projectileEmitter.projectileVelocity.y * directionY;
                     
@@ -67,19 +77,24 @@ class ProjectileEmitterSystem: public System {
         }
         
         void Update(std::unique_ptr<Registry>& registry) {
-            for (auto entity: GetSystemEntity()) {
+            for (auto entity: GetSystemEntity()) 
+            {
                 auto& projectileEmitter = entity.GetComponent<ProjectileEmitterComponent>();
                 const auto transform = entity.GetComponent<TransformComponent>();
 
                 // If emission frequency is zero, bypass re-emission logic
-                if (projectileEmitter.repeatFrequency == 0) {
+                if (projectileEmitter.repeatFrequency == 0) 
+                {
                     continue;
                 }
 
                 // Check if its time to re-emit a new projectile
-                if (SDL_GetTicks() - projectileEmitter.lastEmissionTime > projectileEmitter.repeatFrequency) {
+                if (SDL_GetTicks() - static_cast<Uint32>(projectileEmitter.lastEmissionTime) > static_cast<Uint32>(projectileEmitter.repeatFrequency))
+                {
                     glm::vec2 projectilePosition = transform.position;
-                    if (entity.HasComponent<SpriteComponent>()) {
+                    
+                    if (entity.HasComponent<SpriteComponent>()) 
+                    {
                         const auto sprite = entity.GetComponent<SpriteComponent>();
                         projectilePosition.x += (transform.scale.x * sprite.width / 2);
                         projectilePosition.y += (transform.scale.y * sprite.height / 2);
